# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Timesheets AI API",
    "description": "API for time tracking, timesheet approvals, and productivity insights",
    "contact": {
      "name": "Cybernauts",
      "email": "Cybernauts301@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/account/deletion/request": {
      "post": {
        "tags": [
          "account-controller"
        ],
        "summary": "Request account deletion",
        "description": "Submits an account deletion request for the authenticated user. The request is sent for administrative review before the account is deleted.",
        "operationId": "requestDeletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Account deletion request submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "Account deletion request submitted successfully. An admin will review your request."
                }
              }
            }
          },
          "400": {
            "description": "Invalid account deletion request",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Authentication is required to request account deletion",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "redirectUrl": {
            "type": "string"
          }
        }
      },
...
      "Request": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "reason"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}